+++ packages/ui/src/components/button-group/components/button-group-button/button-group-button.component.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { mergeProps } from '@react-aria/utils';
 import React, { useContext, useRef } from 'react';
 import { useFocusRing, useToggleButtonGroupItem } from 'react-aria';
 
@@ -32,8 +33,7 @@ export function ButtonGroupButton({ className, ...props }: ButtonGroupButtonProp
 
   return (
     <Button
-      {...buttonProps}
-      {...focusProps}
+      {...mergeProps(buttonProps, focusProps)}
       className={styles.button({ className })}
       soft={!isSelected}
       data-pressed={isPressed}