+++ packages/ui/src/components/multi-select/components/multi-select-dropdown/multi-select-dropdown.types.ts
@@ -1,0 +1,6 @@
+import { Dispatch, SetStateAction } from 'react';
+import { AriaListBoxOptions } from 'react-aria';
+
+export type MultiSelectDropdownProps<T> = {
+  setFilterText: Dispatch<SetStateAction<string>>;
+} & AriaListBoxOptions<T>;