+++ packages/ui/src/components/multi-select/components/multi-select-list-box-trigger/multi-select-list-box-trigger.types.ts
@@ -1,0 +1,21 @@
+import { AriaButtonProps } from 'react-aria';
+import { ListProps } from 'react-stately';
+import { VariantProps } from 'tailwind-variants';
+
+import { ResponsiveVariants } from 'src/types/responsive-variants.types.js';
+
+import { MultiSelectProps } from '../../multi-select.types.js';
+
+import { styles as triggerStyles } from './multi-select-list-box-trigger.styles.js';
+
+type Variants = VariantProps<typeof triggerStyles>;
+
+export type MultiSelectSize = ResponsiveVariants<Variants['size']>;
+
+export type MultiSelectListBoxTriggerProps<T> = {
+  id?: string;
+  placeholder: string;
+  selectedKeys?: ListProps<T>['selectedKeys'];
+  showSingleSectionTitle?: MultiSelectProps<T>['showSingleSectionTitle'];
+  triggerProps: AriaButtonProps<'button'>;
+};