+++ packages/ui/src/components/tooltip/components/tooltip-content/tooltip-content.types.ts
@@ -3,4 +3,5 @@ import { ReactNode } from 'react';
 export type TooltipContentProps = {
   children?: ReactNode;
   id: string;
+  position?: 'top' | 'bottom';
 };